rename surface.py->class_initialization.py
